% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est app_package.r
\name{Xu2010}
\alias{Xu2010}
\title{A Cat Function}
\usage{
Xu2010(T1, T2, d2, int_theta, tol = tol, step)
}
\arguments{
\item{T1}{observed mediator event time (vector)}

\item{T2}{observed terminal event time (vector)}

\item{d2}{1 for terminal event occured 0 for censored (vector)}

\item{int_theta}{initial value (>0) for theta used for iteration}

\item{tol}{maximum tolerance of change during the iteration}

\item{step}{maximum  number of the iteration}
}
\description{
Estimating the parameters of the frailty model
}
\examples{
data=meta.gen(500,theta_0=0.5,theta_1=0.5,L1=1,L2=1,L3=1,b01=0,b02=0,b03=0,cc=2,dd="uniform")
ans=Xu2010(data$X1,data$X2,data$D, int_theta=1 ,tol=0.01,step=50,FIG="FALSE")
ans
}
\keyword{Xu2010}
          
